requirements.txt
langchain>=0.2.11
langchain-community>=0.2.10
langchain-google-genai>=2.0.0
google-generativeai>=0.7.2
chromadb>=0.5.3
pypdf>=4.2.0
python-dotenv>=1.0.1
numpy>=1.26.4


